OVU-153 Corrects the save & close vs. close functionality
